feat: testing cli jira
