+++ pyproject.toml
@@ -16,11 +16,14 @@ repository = "https://github.com/FLC-Data/flc-core-data"
   pre-commit = "^2.17.0"
   requests = "^2.27.1"
 
+[tool.poetry.group.dev.dependencies]
+pyment = "^0.3.3"
+
   [tool.poetry.dev-dependencies]
   flake8 = "3.9.2"
 
   [tool.poetry.scripts]
-  jira-script = "flc_core_data.hello_world:cli"
+  jira-issue-script = "flc_core_data.jira.issue_description:cli"
 
 [build-system]
 requires = ["poetry-core"]
